+++ django_spire/notification/app/views/page_views.py
@@ -10,12 +10,20 @@ if TYPE_CHECKING:
     from django.core.handlers.wsgi import WSGIRequest
     from django.template.response import TemplateResponse
 
+
 def app_notification_list_view(request: WSGIRequest) -> TemplateResponse:
-    app_notifification_list = AppNotification.objects.by_user(request.user).order_by('-created_datetime')
+    app_notification_list = (
+        AppNotification.objects.active()
+        .is_sent()
+        .annotate_is_viewed_by_user(request.user)
+        .select_related("notification")
+        .distinct()
+        .order_by('-notification__sent_datetime')
+    )
 
     return portal_views.list_view(
         request,
-        context_data={'notification_list': app_notifification_list},
+        context_data={'notification_list': app_notification_list},
         model=AppNotification,
         template='django_spire/notification/app/page/list_page.html'
     )