+++ django_spire/notification/email/exceptions.py
@@ -1,0 +1,5 @@
+from django_spire.exceptions import DjangoSpireException
+
+
+class EmailNotificationException(DjangoSpireException):
+    pass