+++ django_spire/notification/sms/admin.py
@@ -1,3 +1,10 @@
 from django.contrib import admin
 
-# Register your models here.
+from django_spire.notification.sms.models import SmsNotification
+
+
+@admin.register(SmsNotification)
+class EmailNotificationAdmin(admin.ModelAdmin):
+    list_display = (
+        'id', 'notification', 'to_phone_number'
+    )