+++ django_spire/notification/sms/querysets.py
@@ -1,0 +1,5 @@
+from django_spire.history.querysets import HistoryQuerySet
+
+
+class SmsNotificationQuerySet(HistoryQuerySet):
+    pass