+++ django_spire/notification/urls.py
@@ -4,5 +4,5 @@ from django.urls import path, include
 app_name = 'notification'
 
 urlpatterns = [
-    path('app/', include('django_spire.notification.app.urls', namespace='group')),
+    path('app/', include('django_spire.notification.app.urls', namespace='app')),
 ]